Eval of exogen parameter fixed (remote)
The varest objects hold local variable names for exogen in the call
elements. These local variable names are possibly not in scope after
calling the prediction function, throwing an error in the eval of this
call$exogen clause. The call element is fixed by calling eval on the
local variable prior to returning the object.