Call element in varest object fixed for exogen
Problem: when the VAR function is called from another function with a variable as exogen parameter from the outermost scope, i.e. VAR(y, type="none", exogen=some.local.variable), the call is stored with "some.local.variable" literally as exogen parameter. Calling a different function, e.g. predict(), tries to eval "some.local.variable" which is possibly not in scope anymore resulting in an error.

Solution: the exogen parameter in the call is stored as a value instead of the variable name in the call part of the varest object.